requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.4
python-socketio==5.8.0
SpeechRecognition==3.10.0
pyaudio==0.2.11
google-cloud-speech==2.21.0
google-generativeai==0.3.1
transformers==4.33.2
torch==2.0.1
torchvision==0.15.2
openai==1.109.1
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.0.1
numpy==1.24.3
pymongo==4.5.0
gunicorn==21.2.0
boto3==1.28.57
diffusers==0.21.2
accelerate==0.23.0